# specs/004-backtest-evaluation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Backtest and Evaluation API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /evaluation-results:
    post:
      operationId: evaluateBacktestResult
      summary: Calculate metrics and apply exact versioned policies
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEvaluationRequest' }
      responses:
        '201':
          description: Evaluation created or resolved idempotently
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvaluationResultEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /evaluation-results/{evaluationResultId}:
    get:
      operationId: getEvaluationResult
      parameters: [{ $ref: '#/components/parameters/EvaluationResultId' }]
      responses:
        '200':
          description: Evaluation returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvaluationResultEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
  /evaluation-comparisons:
    post:
      operationId: compareEvaluationResults
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ComparisonRequest' }
      responses:
        '200':
          description: Comparison returned with compatibility findings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ComparisonEnvelope' }
        '422': { $ref: '#/components/responses/ValidationError' }
components:
  parameters:
...
    EvaluationResultId: { name: evaluationResultId, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
...
    Decimal: { type: string, pattern: '^-?[0-9]+(\.[0-9]+)?$' }
...
    MetricSet:
      type: object
      required: [totalReturn, winRate, maxDrawdown, numberOfTrades, profitFactor, sharpeRatio]
      properties:
        totalReturn: { $ref: '#/components/schemas/Decimal' }
        winRate: { $ref: '#/components/schemas/Decimal' }
        maxDrawdown: { $ref: '#/components/schemas/Decimal' }
        numberOfTrades: { type: integer, minimum: 0 }
        profitFactor: { oneOf: [{ $ref: '#/components/schemas/Decimal' }, { type: 'null' }] }
        sharpeRatio: { oneOf: [{ $ref: '#/components/schemas/Decimal' }, { type: 'null' }] }
    CreateEvaluationRequest:
      type: object
      required: [backtestResultId, evaluationPolicyId, evaluationPolicyVersion, scoringPolicyId, scoringPolicyVersion]
      properties:
        backtestResultId: { type: string, format: uuid }
        evaluationPolicyId: { type: string }
        evaluationPolicyVersion: { type: string }
        scoringPolicyId: { type: string }
        scoringPolicyVersion: { type: string }
    EvaluationResult:
      type: object
      required: [id, backtestResultId, jobId, runId, strategyId, strategyVersion, datasetId, pair, timeframe, startTime, endTime, executionConfig, metrics, score, eligible, exclusionReasons, evaluationPolicyId, evaluationPolicyVersion, scoringPolicyId, scoringPolicyVersion, evaluatedAt, analysisType, disclaimer]
      properties:
        id: { type: string, format: uuid }
        backtestResultId: { type: string, format: uuid }
        jobId: { type: string, format: uuid }
        runId: { type: string, format: uuid }
        strategyId: { type: string }
        strategyVersion: { type: string }
        datasetId: { type: string, format: uuid }
        pair: { type: string }
        timeframe: { type: string }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        executionConfig: { type: object, additionalProperties: true }
        metrics: { $ref: '#/components/schemas/MetricSet' }
        score: { $ref: '#/components/schemas/Decimal' }
        eligible: { type: boolean }
        exclusionReasons: { type: array, items: { type: string } }
        evaluationPolicyId: { type: string }
        evaluationPolicyVersion: { type: string }
        scoringPolicyId: { type: string }
        scoringPolicyVersion: { type: string }
        evaluatedAt: { type: string, format: date-time }
        analysisType: { type: string, const: HISTORICAL_SIMULATION }
        disclaimer: { type: string, minLength: 1 }
    ComparisonRequest:
      type: object
      required: [evaluationResultIds, mode]
      properties:
        evaluationResultIds: { type: array, minItems: 2, maxItems: 20, uniqueItems: true, items: { type: string, format: uuid } }
        mode: { type: string, enum: [STRICT, CONTEXTUAL] }
    Comparison:
      type: object
      required: [compatible, differences, results]
      properties:
        compatible: { type: boolean }
        differences: { type: array, items: { type: object, required: [dimension, values], properties: { dimension: { type: string }, values: { type: array, items: {} } } } }
        results: { type: array, items: { $ref: '#/components/schemas/EvaluationResult' } }
...
    ResponseMeta:
      type: object
      required: [success, message, timestamp, requestId]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        timestamp: { type: string, format: date-time }
        requestId: { type: string }
...
    EvaluationResultEnvelope: { allOf: [{ $ref: '#/components/schemas/ResponseMeta' }, { type: object, required: [data], properties: { data: { $ref: '#/components/schemas/EvaluationResult' } } }] }
    ComparisonEnvelope: { allOf: [{ $ref: '#/components/schemas/ResponseMeta' }, { type: object, required: [data], properties: { data: { $ref: '#/components/schemas/Comparison' } } }] }
...
    ErrorEnvelope:
      type: object
      required: [success, message, error, timestamp, requestId]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        error:
          type: object
          required: [code, retryable]
          properties:
            code: { type: string }
            retryable: { type: boolean }
            details: { type: [object, 'null'], additionalProperties: true }
        timestamp: { type: string, format: date-time }
        requestId: { type: string }
  responses:
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    ValidationError: { description: Semantic validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...